+++ engine/spec/lib/citizens_advice_components/elements/text_input_spec.rb
@@ -69,6 +69,14 @@ RSpec.describe CitizensAdviceComponents::Elements::TextInput do
       end
     end
 
+    context "with 'page_heading' parameter" do
+      it "passes hint to the text input component" do
+        builder.cads_text_field(:name, page_heading: true)
+
+        expect(component).to have_received(:new).with(hash_including(options: hash_including(page_heading: true)))
+      end
+    end
+
     context "when there is a validation error" do
       it "sets 'error_message'" do
         model.errors.add(:name, :example, message: "has an example error")